requirements.txt
Django==3.1
Pillow==8.1
psycopg2-binary==2.8.6
djangorestframework
python-decouple
django-ckeditor
pytils
django-filter
drf-yasg